feat: add schema checker and expand ORM model coverage
Add SchemaChecker to validate SQLAlchemy models against Access database
schema with diff report generation. Expand model coverage with new energy
and pipenet modules, and fix type mismatches in existing models based on
schema validation results.

- Add schema_checker.py with schema comparison and markdown report generation
- Add energy.py models (EnergyDevice, EnergyHotwater, EnergyLiftEscalator, EnergyPumpGroup)
- Add pipenet.py models (DnFk, DnJoint, DnPoint, HacnetOption, PnStart, PnTerm)
- Add AccdbReader methods for schema introspection
- Add check_schema CLI command
- Fix column types in building, construction, hvac, misc models
- Export new models from __init__.py

+++ src/database/models/environment.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import Float, ForeignKey, Integer, String
+from sqlalchemy import Float, ForeignKey, Integer, SmallInteger, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from ._base import Base
@@ -74,7 +74,7 @@ class ClimateData(Base):
     )
     t_sky: Mapped[float | None] = mapped_column(Float, comment='Sky temperature (°C)')
     ws: Mapped[float | None] = mapped_column(Float, comment='Wind speed (m/s)')
-    wd: Mapped[float | None] = mapped_column(Float, comment='Wind direction (°)')
+    wd: Mapped[int | None] = mapped_column(SmallInteger, comment='Wind direction (°)')
     b: Mapped[float | None] = mapped_column(Float, comment='Atmospheric pressure (Pa)')
